requirements.txt
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
flask-json-schema==0.0.5
PyJWT==2.6.0
uuid==1.30